Extract voice logic into module
